refactor: move types and code to corresponding folder
