+++ src/graduate-types/index.ts
@@ -1,0 +1,2 @@
+export * from "./common";
+export * from "./major2";
