+++ src/parse/types.ts
@@ -1,0 +1,4 @@
+import { TypedCatalogEntry } from "../classify";
+import { Major2 } from "../graduate-types";
+
+export type ParsedCatalogEntry = TypedCatalogEntry & { parsed: Major2 };
